tests/requirements.txt
pytest>=8.0.0
pytest-asyncio
pytest-playwright
httpx
sqlalchemy
pydantic
bcrypt
faker
